Bug fixes
- BitmaskStackpanel with option default value (class & race needs -1 to
allow all)
- Bmsp displays in preview using event handler
- long instead of uint

+++ TrinityCreator/Converters/ValueToVisibilityConverter.cs
@@ -28,7 +28,7 @@ namespace TrinityCreator
                 flag = true;
             else if (value is int && (int)value != 0)
                 flag = true;
-            else if (value is uint && (uint)value != 0)
+            else if (value is long && (long)value != 0)
                 flag = true;
             else if (value is double && (double)value != 0 && !double.IsNaN((double)value) && !double.IsInfinity((double)value))
                 flag = true;
